+++ src/Symfony/Component/Messenger/Bridge/Amqp/Transport/AmqpStamp.php
@@ -17,6 +17,7 @@ use Symfony\Component\Messenger\Stamp\NonSendableStampInterface;
  * @author Guillaume Gammelin <ggammelin@gmail.com>
  * @author Samuel Roze <samuel.roze@gmail.com>
  */
+#[\Attribute(\Attribute::TARGET_CLASS | \Attribute::IS_REPEATABLE)]
 final class AmqpStamp implements NonSendableStampInterface
 {
     private ?string $routingKey;