+++ src/Symfony/Component/Messenger/Stamp/SerializerStamp.php
@@ -14,6 +14,7 @@ namespace Symfony\Component\Messenger\Stamp;
 /**
  * @author Maxime Steinhausser <maxime.steinhausser@gmail.com>
  */
+#[\Attribute(\Attribute::TARGET_CLASS | \Attribute::IS_REPEATABLE)]
 final class SerializerStamp implements StampInterface
 {
     private array $context;